MAINT tiny harness API a little bit
